+++ tasks/cmg_task.h
@@ -62,7 +62,7 @@ public:
       m_path = state_.m_path;
     }
 
-    void do_action(int action) { m_mode_idx = action; }
+    void do_action(unsigned long int action) { m_mode_idx = action; }
 
     State &operator=(const State &state_) {
       this->m_pose = state_.m_pose;
@@ -81,7 +81,7 @@ public:
     int t_max = -1; // maximum time step this can reach
     State2() {}
     State2(int t, int idx) : timestep(t), finger_index(idx) {}
-    void do_action(int action) {
+    void do_action(unsigned long int action) {
       this->finger_index = action;
       this->timestep++;
     }
@@ -175,12 +175,12 @@ public:
   double total_finger_change_ratio(const std::vector<State2> &path);
   double evaluate_path(const std::vector<State2> &path);
 
-  double estimate_next_state_value(const State2 &state, int action) {
+  double estimate_next_state_value(const State2 &state, unsigned long int action) {
     // return 0.0 for now, can use neural networks to estimate values
     return 0.0;
   }
 
-  double action_heuristics_level2(int action_idx, const State2 &state,
+  double action_heuristics_level2(unsigned long int action_idx, const State2 &state,
                                   const State2 &pre_state) {
     // return the heuristics of an action in level2, this can be hand designed
     // or learned todo: improve this heuristics
@@ -265,7 +265,7 @@ public:
     return timestep * this->n_finger_combinations + finger_idx;
   }
 
-  void do_action(State2 &state, int action) {
+  void do_action(State2 &state, unsigned long int action) {
     state.timestep = action / this->n_finger_combinations;
     state.finger_index = action % this->n_finger_combinations;
   }
